refactor: layout 분리
