+++ src/pages/_app.tsx
@@ -6,6 +6,7 @@ import { RecoilRoot } from 'recoil';
 import ToastSection from '~/components/common/ToastSection';
 import GlobalStyle from '~/styles/GlobalStyle';
 import CustomTheme from '~/styles/Theme';
+import Layout from '~/components/common/Layout';
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
@@ -20,14 +21,3 @@ export default function App({ Component, pageProps }: AppProps) {
     </RecoilRoot>
   );
 }
-
-function Layout({ children }: PropsWithChildren<{}>) {
-  return <div css={layoutCss}>{children}</div>;
-}
-
-const layoutCss = (theme: Theme) => css`
-  max-width: ${theme.size.maxWidth};
-  width: 100%;
-  margin: 0 auto;
-  padding: ${theme.size.layoutPadding};
-`;
